requirements_enhanced.txt
streamlit>=1.35.0
pydantic>=2.7.0
openai>=1.30.0
python-dotenv>=1.0.1
tenacity>=8.3.0
plotly>=5.15.0
pandas>=2.0.0